+++ src/gmprocess/io/asdf/flatfile.py
@@ -74,9 +74,7 @@ class Flatfile(object):
         """
         if hasattr(self.workspace, "config"):
             config = self.workspace.config
-            default_config_file = (
-                constants.DATA_DIR / constants.CONFIG_FILE_PRODUCTION
-            )
+            default_config_file = constants.DATA_DIR / constants.CONFIG_FILE_PRODUCTION
             with open(default_config_file, "r", encoding="utf-8") as f:
                 yaml = YAML()
                 yaml.preserve_quotes = True
@@ -87,9 +85,7 @@ class Flatfile(object):
         any_trace_failures = config["check_stream"]["any_trace_failures"]
         use_array = not config["read"]["use_streamcollection"]
 
-        wmc = WaveformMetricCollection.from_workspace(
-            self.workspace, self.label
-        )
+        wmc = WaveformMetricCollection.from_workspace(self.workspace, self.label)
         smc = StationMetricCollection.from_workspace(self.workspace)
         for wml, stream_meta, station_metrics in zip(
             wmc.waveform_metrics, wmc.stream_metadata, smc.station_metrics
@@ -98,9 +94,7 @@ class Flatfile(object):
                 continue
 
             self._station_metrics = station_metrics
-            passed_traces = [
-                tr_met for tr_met in stream_meta if tr_met["passed"]
-            ]
+            passed_traces = [tr_met for tr_met in stream_meta if tr_met["passed"]]
 
             n_passed = len(passed_traces)
             for tr in passed_traces:
@@ -286,9 +280,7 @@ class Flatfile(object):
                 imt_cols = [x for x in imt_cols if x not in fas_cols]
             imt_cols.sort(key=_natural_keys)
 
-            non_imt_cols = [
-                col for col in table.columns if col not in self.imtlist
-            ]
+            non_imt_cols = [col for col in table.columns if col not in self.imtlist]
             table = table[non_imt_cols + imt_cols]
             self.imc_tables[key] = table
             readme_dict = {}
@@ -298,17 +290,19 @@ class Flatfile(object):
                 else:
                     imt = col.upper()
                     if imt.startswith("SA"):
-                        readme_dict["SA(X)"] = flat_const.FLATFILE_IMT_COLUMNS[
-                            "SA(X)"
-                        ]
+                        readme_dict["SA(X)"] = flat_const.FLATFILE_IMT_COLUMNS["SA(X)"]
+                    elif imt.startswith("SV"):
+                        readme_dict["SV(X)"] = flat_const.FLATFILE_IMT_COLUMNS["SV(X)"]
+                    elif imt.startswith("SD"):
+                        readme_dict["SD(X)"] = flat_const.FLATFILE_IMT_COLUMNS["SD(X)"]
                     elif imt.startswith("FAS"):
                         readme_dict["FAS(X)"] = flat_const.FLATFILE_IMT_COLUMNS[
                             "FAS(X)"
                         ]
                     elif imt.startswith("DURATION"):
-                        readme_dict["DURATIONp-q"] = (
-                            flat_const.FLATFILE_IMT_COLUMNS["DURATIONp-q"]
-                        )
+                        readme_dict["DURATIONp-q"] = flat_const.FLATFILE_IMT_COLUMNS[
+                            "DURATIONp-q"
+                        ]
                     else:
                         readme_dict[imt] = flat_const.FLATFILE_IMT_COLUMNS[imt]
             df_readme = pd.DataFrame.from_dict(readme_dict, orient="index")
@@ -355,9 +349,7 @@ class Flatfile(object):
                         fit_dict["StationLongitude"] = coords.longitude
                         fit_dict["StationElevation"] = coords.elevation
                         if trace.has_parameter("corner_frequencies"):
-                            freq_dict = trace.get_parameter(
-                                "corner_frequencies"
-                            )
+                            freq_dict = trace.get_parameter("corner_frequencies")
                             fit_dict["fmin"] = freq_dict["highpass"]
                             fit_dict["fmax"] = freq_dict["lowpass"]
                         fit_table.append(fit_dict)
@@ -370,9 +362,7 @@ class Flatfile(object):
         # Ensure that the DataFrame columns are ordered correctly
         fit_df = fit_df[flat_const.FIT_SPECTRA_COLUMNS.keys()]
 
-        readme = pd.DataFrame.from_dict(
-            flat_const.FIT_SPECTRA_COLUMNS, orient="index"
-        )
+        readme = pd.DataFrame.from_dict(flat_const.FIT_SPECTRA_COLUMNS, orient="index")
         readme.reset_index(level=0, inplace=True)
         readme.columns = ["Column header", "Description"]
 
@@ -472,10 +462,7 @@ def _natural_keys(text):
     Helper function for sorting IMT list. This is needed because using the
     plain .sort() will put SA(10.0) before SA(2.0), for example.
     """
-    return [
-        _atof(c)
-        for c in re.split(r"[+-]?([0-9]+(?:[.][0-9]*)?|[.][0-9]+)", text)
-    ]
+    return [_atof(c) for c in re.split(r"[+-]?([0-9]+(?:[.][0-9]*)?|[.][0-9]+)", text)]
 
 
 def _atof(text):