+++ tests/gmprocess/metrics/waveform_metric_calculator_test.py
@@ -94,6 +94,7 @@ def test_metric_calculator(load_data_us1000778i, config):
             }
         },
     }
+
     conf["metrics"] = metric_config
 
     stream = streams[0]
@@ -116,21 +117,37 @@ def test_metric_calculator(load_data_us1000778i, config):
     np.testing.assert_allclose(pgv["RotD(percentile=100.0)"], 33.284090745430966)
 
     sa1 = wml.select("SA", period=1.0, damping=0.05)[0].values
-    np.testing.assert_allclose(sa1["Channels(component=H1)"], 42.135839214645564)
+
+    # if sa1["RotD(percentile=50.0)"] < 42: # conditional breakpoint
+    #     breakpoint()
+
+    np.testing.assert_allclose(sa1["Channels(component=H1)"], 42.135839)
     np.testing.assert_allclose(sa1["Channels(component=H2)"], 41.891393940234074)
     np.testing.assert_allclose(sa1["Channels(component=Z)"], 12.789046181803089)
     np.testing.assert_allclose(sa1["RotD(percentile=50.0)"], 42.13583921464555)
     np.testing.assert_allclose(sa1["RotD(percentile=100.0)"], 47.910999673071736)
 
     sv1 = wml.select("SV", period=1.0, damping=0.05)[0].values
+
+    # if sv1["RotD(percentile=50.0)"] < 60: # conditional breakpoint
+    #     breakpoint()
+
     np.testing.assert_allclose(sv1["Channels(component=H1)"], 63.02799719)
     np.testing.assert_allclose(sv1["Channels(component=H2)"], 62.15551237)
     np.testing.assert_allclose(sv1["Channels(component=Z)"], 20.49380808)
+    np.testing.assert_allclose(sv1["RotD(percentile=50.0)"], 64.763622)
+    np.testing.assert_allclose(sv1["RotD(percentile=100.0)"], 70.624906)
 
     sd1 = wml.select("SD", period=1.0, damping=0.05)[0].values
+
+    # if sd1["RotD(percentile=50.0)"] > 10.5: # conditional breakpoint
+    #     breakpoint()
+
     np.testing.assert_allclose(sd1["Channels(component=H1)"], 10.42514)
     np.testing.assert_allclose(sd1["Channels(component=H2)"], 10.35811174)
     np.testing.assert_allclose(sd1["Channels(component=Z)"], 3.15898147)
+    np.testing.assert_allclose(sd1["RotD(percentile=50.0)"], 10.42514)
+    np.testing.assert_allclose(sd1["RotD(percentile=100.0)"], 11.846505)
 
     fas = wml.select("FAS")[0].values
     fas_qm = fas["QuadraticMean()"]